+++ docs/source/changelog.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--
-
 ## [unreleased]
 
 ### Added
+
+* [ENH] allow "+" in entity labels #680 by @Remi-Gau
 
 ### Changed
 
@@ -20,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
--->
 
 ## [v0.3.0] - 2024-09-11